ar-building-v2/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
bcrypt==4.2.1
python-jose==3.3.0
python-multipart==0.0.18
aiohttp==3.11.10
numpy>=2.0.0
aiodns>=3.0.0